Reduce complexity of DMR packet handler
